% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_calc.R
\name{abundance_calc}
\alias{abundance_calc}
\title{Calculates Abundance}
\usage{
abundance_calc(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' usually created by \code{\link{as.seqData}}.}
}
\value{
An object of class abunRes (also a data.frame) containing the abundance value for every sample in the data object.
}
\description{
This function calculates the total number of features seen in each sample.
}
\details{
Abundance is the total number of features (individuals) seen in each sample.
}
\examples{
library(mintJansson)
data(rRNA_data)
rRNA_abundance <- abundance_calc(omicsData = rRNA_data)
rRNA_abundance
summary(rRNA_abundance)
plot(rRNA_abundance)

}
\author{
Allison Thompson
}
