% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaDiv_calc.R
\name{alphaDiv_calc}
\alias{alphaDiv_calc}
\title{Calculates Alpha Diversity Indices}
\usage{
alphaDiv_calc(omicsData, index = c("shannon", "simpson", "invsimpson"))
}
\arguments{
\item{omicsData}{an object of the class 'gDNAdata', 'cDNAdata', or 'rRNAdata' usually created by \code{\link{as.gDNAdata}}, \code{\link{as.cDNAdata}}, or \code{\link{as.rRNAdata}}, respectively.}

\item{index}{a character vector stating which of the calculations to perform - "shannon" for Shannon's diversity index, "simpson" for Simpson's diversity index, and/or "invsimpson" for the inverse Simpson's diversity index. Default is to perform all 3 calculations.}
}
\value{
An object of class alphaRes (also a data.frame) containing the diversity value(s) for every sample in the data object.
}
\description{
This function calculates Shannon's, Simpson's, and/or Inverse Simpson's alpha diversity indices.
}
\details{
Alpha diversity is calculated for each sample in the data, using Shannon's diversity index (Shannon's H), Simpson's diversity index (Simpson's D), and/or inverse Simpson's diversity index.

Shannon or Shannon–Weaver (or Shannon–Wiener) index is defined as H = -sum(p_i*log(p_i)), where p_i is the proportional abundance of species i.

Both variants of Simpson's index are based on D = sum(p_i^2). Choice simpson returns 1-D and invsimpson returns 1/D.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_diversity <- alphaDiv_calc(omicsData = rRNA_data)
rRNA_diversity
summary(rRNA_diversity)
plot(rRNA_diversity)
}

}
\author{
Allison Thompson
}
