% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{applyFilt_worker}
\alias{applyFilt_worker}
\title{Remove items that need to be filtered out}
\usage{
applyFilt_worker(filter_object, omicsData)
}
\arguments{
\item{filter_object}{a list created by the functions above}

\item{omicsData}{an object of the class \code{cDNAdata}, \code{gDNAdata}, or \code{rRNAdata} usually created by \code{\link{as.cDNAdata}}, \code{\link{as.gDNAdata}}, \code{\link{as.rRNAdata}}, respectively.}
}
\value{
list similar to as.mintRData object
}
\description{
This function removes filtered objects
}
\author{
Kelly Stratton, Lisa Bramer, Allison Thompson
}
