% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_based_filter.R
\name{count_based_filter}
\alias{count_based_filter}
\title{Count filter object}
\usage{
count_based_filter(omicsData, fn = "sum")
}
\arguments{
\item{omicsData}{An object of one of the classes "seqData"}

\item{fn}{Specify "mean" to use the mean count of each OTU, "percent" to use
mean counts lower than a certain percent, "max" to use the max count across
all samples, "sum" to use the total count of each OTU, "nonmiss" to use
presence/absence counts, or "ka" to use k over a filtering (need at least
k counts of OTUs seen in at least a samples).}
}
\value{
An object of class countFilter (also a data.frame) that contains the
  molecule identifier and the mean/percent/max/sum/nonmissing count across
  all samples.
}
\description{
This function returns a countFilter object, performing all of the
calculations needed to filter the data based off a specified function and
limit
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
omicsData <- rRNA_data

#Find mean count of OTUs
mean_lim <- count_based_filter(omicsData, fn="mean")
head(mean_lim)
summary(mean_lim)
plot(mean_lim)

#Find percentage of each OTU
perc_lim <- count_based_filter(omicsData, fn="percent")
head(perc_lim)
summary(perc_lim)
plot(perc_lim)

#Find maximum count of OTUs
max_lim <- count_based_filter(omicsData, fn="max")
head(max_lim)
summary(max_lim)
plot(max_lim)

#Find total count of OTUs
sum_lim <- count_based_filter(omicsData, fn="sum")
head(sum_lim)
summary(sum_lim)
plot(sum_lim)

#Find number of nonmissing OTUs
nonmiss_lim <- count_based_filter(omicsData, fn="nonmiss")
head(nonmiss_lim)
summary(nonmiss_lim)
plot(nonmiss_lim)

#Find order of samples for k/a filtering
ka_lim <- count_based_filter(omicsData, fn="ka")
head(ka_lim)
summary(ka_lim)
plot(ka_lim)
}

}
\references{
Arumugam, Manimozhiyan, et al. "Enterotypes of the human gut
microbiome." nature 473.7346 (2011): 174-180.

https://bioinformatics.oxfordjournals.org/content/early/2013/07/15/bioinformatics.btt350.full
}
\author{
Allison Thompson, Bryan Stanfill
}
