% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSTAT.R
\name{countSTAT}
\alias{countSTAT}
\title{Statistical tests for count data}
\usage{
countSTAT(omicsData, norm_factors = NULL, comparisons, control = NULL, test,
  pval_adjust = "BH", pval_thresh = 0.05)
}
\arguments{
\item{omicsData}{an object of the class 'gDNAdata', 'cDNAdata', or 'rRNAdata' usually created by \code{\link{as.gDNAdata}}, \code{\link{as.cDNAdata}},or \code{\link{as.rRNAdata}}, respectively.}

\item{norm_factors}{Named vector of normalization parameters to put into DESeq2 or edgeR. If NULL, will use DESeq2's or edgeR's inbuilt normalization. Default is NULL.}

\item{comparisons}{dictates which pairwise comparisons to make. 'all' will give the results for all pairwise comparisons, 'control' will give the results for all comparisons against a specified control group, or a list of specific comparisons to be made or terms for a paired test can be given.}

\item{control}{only necessary when performing comparisons against control, name of the control group. default is NULL.}

\item{test}{names of which differential expression test(s) to use, options are "dw" (DESeq2 analysis with Wald test), "dl" (DESeq2 analysis with LRT test), "dp" (DESeq2 analysis on paired data), "eq" (edgeR analysis with QCML test), "el" (edgeR analysis with LRT test), "ef" (edgeR analysis with QL F-test), and/or "ep" (edgeR analysis on paired data). At this point, cannot perform a paired analysis at the same time as the other tests.}

\item{pval_adjust}{Name of which multiple comparisons adjustment to use. Options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "FDR", and "none". See ?p.adjust for more information. The default is "BH".}

\item{pval_thresh}{P-value threshold for significance. The default is 0.05.}
}
\value{
A results object containing the log2 fold change (logFC), log-average concentration/abundance (logCPM), likelihood ratio (LR), exact p-value for differential expression using the negative binomial model (PValue), and the p-value adjusted for multiple testing (FDR) for every pairwise comparison and every feature.
}
\description{
Differential expression tests for count data
}
\details{
Perform pairwise differential abundance analysis of omicsData using DESeq2 and/or edgeR.
}
\examples{
\dontrun{
library(mintJansson)
data(cDNA_hiseq_data)
mycdnadata <- group_designation(omicsData = cDNA_hiseq_data, main_effects = c("treatment"), time_course=NULL)
mycdnadata_norm <- normalize_data(omicsData = mycdnadata, norm_fn = "percentile")
mycdnadata_dw_results <- countSTAT(omicsData = mycdnadata_norm, comparisons = list(c("Neg","Plus")), control = NULL, test = "dw", pval_adjust = "none", pval_thresh = 0.05)
mycdnadata_all_results <- countSTAT(omicsData = mycdnadata_norm, comparisons = "all", control = NULL, test = c("dw","eq","el","ef"), pval_adjust = "BH", pval_thresh = 0.05)
}

}
\author{
Allison Thompson
}
