% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{CSS_Norm}
\alias{CSS_Norm}
\title{Cumulative sum scaling normalization of count data}
\usage{
CSS_Norm(e_data, edata_id, q = 0.75, qg = "median")
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}

\item{q}{a number to indicate which quantile to normalize with. Default is 0.75.}

\item{qg}{a number to use for scaling all samples. Default is 1000 (as in reference). Can also specify "median" to use the median value scaling value across all samples.}
}
\value{
List containing 3 elements: norm_data is a data.frame with same structure as e_data that contains the quantile-normalized data, location_param is NULL, and scale_param is a numeric vector containing, for every sample, the value of the sample at the designated quantile (q) divided by the value of the global quantile (q).
}
\description{
The method normalizes count data by cumulative sum scaling, using a specified quantile (e.g., 75th quantile)
}
\details{
Count data is normalized by a given quantile, dividing by the sum of all values up to and including the given quantile of the sample and multiplying by the given scaling value (either 1000 or the median scaling value across all samples).
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_CSS <- CSS_Norm(e_data = rRNA_data$e_data, edata_id = attr(rRNA_data, "cnames")$edata_cname)
norm_factors <- attr(rRNA_CSS,"data_info")$scale_param
}

}
\references{
Paulson, Joseph N, O Colin Stine, Hector Corrada Bravo, and Mihai Pop. "Differential abundance analysis for microbial marker-gene surveys." Nature Methods. 10.12 (2013)
}
\author{
Allison Thompson, Lisa Bramer
}
