% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_modules.R
\name{detect_modules}
\alias{detect_modules}
\title{Detect modules for Network}
\usage{
detect_modules(netGraph, cluster = "louvain", cutoff = 5)
}
\arguments{
\item{netGraph}{an object of class 'networkGraph', created by \code{link{pmartRseq_igraph}}}

\item{cluster}{which clustering method to use. Must be one of 'edge_betweenness', 'fast_greedy', 'infomap', 'label_prop', 'leading_eigen', 'louvain', 'optimal', 'spinglass', or 'walktrap'. Default is 'louvain'.}

\item{cutoff}{Any modules which come back with a total of less than this number of members will all be grouped together.}
}
\value{
An object of class 'modData', with group designations for every vertex in the graph.
}
\description{
This function detects modules for a network.
}
\details{
A network graph is created for the network(s) that were generated.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
mynetwork <- network_calc(omicsData = rRNA_data)
mygraph <- pmartRseq_igraph(netData = mynetwork, coeff=0.6, pval=NULL, qval=0.05)
mymods <- detect_modules(netGraph = mygraph)
}

}
\author{
Allison Thompson
}
