% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effsp_calc.R
\name{effsp_calc}
\alias{effsp_calc}
\title{Calculates Effective Species}
\usage{
effsp_calc(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' reated by \code{\link{as.seqData}}.}
}
\value{
An object of class effspRes (also a data.frame) containing the number of effective species for every sample in the data object.
}
\description{
This function calculates the number of effective species for every sample in a count dataset
}
\details{
Calculates effective species of count data
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_effsp <- effsp_calc(omicsData = rRNA_data)
rRNA_effsp
summary(rRNA_effsp)
plot(rRNA_effsp)
}

}
\author{
Allison Thompson
}
