% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenness_calc.R
\name{evenness_calc}
\alias{evenness_calc}
\title{Calculates Evenness Indices}
\usage{
evenness_calc(omicsData, index = c("shannon", "simpson"))
}
\arguments{
\item{omicsData}{an object of the class 'seqData' reated by \code{\link{as.seqData}}.}

\item{index}{a character vector stating which of the calculations to perform - "shannon" for Shannon's evenness index and/or "simpson" for Simpson's evenness index. Default is to perform both calculations.}
}
\value{
An object of class evenRes (also a data.frame) containing the evenness value(s) for every sample in the data object.
}
\description{
This function calculates Shannon's, Simpson's, and/or Inverse Simpson's evenness indices.
}
\details{
Evenness is calculated for each sample in the data. Evenness is calculated using Shannon's evenness index or Simpson's evenness index.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_even <- evenness_calc(omicsData = rRNA_data)
rRNA_even
summary(rRNA_even)
plot(rRNA_even)
}

}
\author{
Allison Thompson
}
