% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seqData.R
\name{import_seqData}
\alias{import_seqData}
\title{Import rRNA (16S/ITS/18S), metatranscript, or metagenomic data from a .biom,
.csv, or .txt file. Converts import to a list object to be passed to
\link{as.seqData}.}
\usage{
import_seqData(e_data_filepath, f_data_filepath, e_meta_filepath = NULL)
}
\arguments{
\item{e_data_filepath}{character string specifying the file path to a .csv,
.txt, or .biom expression data file. The expression data file should
contain a feature identifier/name.}

\item{f_data_filepath}{character string specifying the file path to a .csv or
.txt sample information file. Each row must correspond to a sample with one
column giving the unique sample identifiers found in e_data column names
and other columns providing qualitative and/or quantitative traits of each
sample. The sample information should also contain a sample identifier
column.}

\item{e_meta_filepath}{optional character string specifying the file path of
expression meta information. The meta information must contain an
identifier column that matches the identifier column in the expression data
file.}
}
\description{
Import rRNA (16S/ITS/18S), metatranscript, or metagenomic data from a .biom,
.csv, or .txt file. Converts import to a list object to be passed to
\link{as.seqData}.
}
\seealso{
\code{\link{as.seqData}}
}
\author{
Sarah Reehl
}
