% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indsp_calc.R
\name{indsp_calc}
\alias{indsp_calc}
\title{Computes Indicator Species Analysis for Count Data}
\usage{
indsp_calc(omicsData, within = NULL, pval_thresh = 0.05, max_grp = NULL)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}.}

\item{within}{If desired, can run indicator species analysis between one effect and within another effect.}

\item{pval_thresh}{P-value threshold for creating flags for significance. The default is 0.05.}

\item{max_grp}{Integer indicating the maximum number of group combinations to be considered. If NULL, will use n-1 groups. Max is n-1 groups.}
}
\value{
An object of class indspRes (also a data.frame) containing the indicator species analysis of the data.
}
\description{
This function is a wrapper for an indicator species analysis using the package \code{indicspecies}, studying the association between species patterns and combinations of group sites.
}
\details{
This function uses the \code{multipatt} function in the package \code{indicspecies} to create combinations of input clusters and compare each combination with the species in \code{e_data}. Please refer to the \code{multipatt} function in the package \code{indicspecies} for more information.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_data <- group_designation(omicsData = rRNA_data, main_effects = c("site", "treatment"), time_course=NULL)

rRNA_indsp <- indsp_calc(omicsData = rRNA_data, within=NULL, pval_thresh=0.05)
head(rRNA_indsp)
summary(rRNA_indsp)
plot(rRNA_indsp)
}

}
\references{
De Cáceres, M. and Legendre, P. 2009. Associations between species and groups of sites: indices and statistical inference. Ecology 90(12): 3566-3574.
}
\author{
Allison Thompson
}
