% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_calc.R
\name{jaccard_calc}
\alias{jaccard_calc}
\title{Calculates Jaccard Index}
\usage{
jaccard_calc(omicsData, sim = TRUE)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}}

\item{sim}{TRUE/FALSE indicating if you want a value for Jaccard similarity (TRUE) or dissimilarity (FALSE). Default is TRUE.}
}
\value{
An object of class jaccardRes (also a data.frame) containing the Jaccard dis/similarity value for every sample in the data object.
}
\description{
This function calculates Jaccard's dis/similarity coefficient between group replicates.
}
\details{
Jaccard dis/similarity index is calculated across replicates within groups. The index is then averaged for each replicate.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_jaccard <- jaccard_calc(omicsData = rRNA_data, sim = TRUE)
rRNA_jaccard
summary(rRNA_jaccard)
plot(rRNA_jaccard)
}

}
\author{
Allison Thompson
}
