% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{med_scounts_norm}
\alias{med_scounts_norm}
\title{Normalization of count data via DESeq scale factors}
\usage{
med_scounts_norm(e_data, edata_id)
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}
}
\value{
List containing 3 elements: norm_data is a data.frame with same structure as e_data that contains the normalized data, location_param is NULL, scale_param is a numeric vector of DESeq scores.
}
\description{
The method normalizes count data by scaling the raw counts by the median scaled counts.  The normalized factors used by
this procedure are also called DEseq size scores (Anders et al. 2010).  The same normalization technique is used in DESeq2 as well (Love et al. 2014).
}
\details{
Count data is normalized by the median scaled score
}
\examples{
library(mintJansson)
e_data_id <- attr(rRNA_data, "cnames")$edata_cname
deseq_n <- med_scounts_norm(e_data = rRNA_data$e_data, e_data_id)

}
\references{
Anders, Simon, and Wolfgang Huber. "Differential expression analysis for sequence count data." Genome biol 11.10 (2010): R106.
Love, Michael I., Wolfgang Huber, and Simon Anders. "Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2." Genome biology 15.12 (2014): 1-21.
}
\author{
Bryan Stanfill
}
