% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_calc.R
\name{network_calc}
\alias{network_calc}
\title{Calculates Correlations}
\usage{
network_calc(omicsData, type = "spearman", group = FALSE,
  group_var = NULL)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' reated by \code{\link{as.seqData}}.}

\item{type}{Character, which correlation metric to use. Must be one of 'pearson' or 'spearman'. Default is 'spearman'.}

\item{group}{Logical, should correlation analysis be performed on sub-groups of the data. Default is FALSE, will use all of the samples in the data.}

\item{group_var}{Character, if group==TRUE, which variable to use as the grouping variable.}
}
\value{
An object of class corrRes (also a data.frame) containing the correlation values.
}
\description{
This function calculates correlations for use in network analysis.
}
\details{
Correlation is calculated between every pair of features.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
mynetwork <- network_calc(omicsData = rRNA_data)
mynetwork
}

}
\references{

}
\author{
Allison Thompson
}
