% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_indices.R
\name{network_indices}
\alias{network_indices}
\title{Calculate Indices (metrics) for Network}
\usage{
network_indices(netGraph)
}
\arguments{
\item{netGraph}{an object of class 'networkGraph', created by \code{link{pmartRseq_igraph}}}
}
\value{
A list containing 1) Metrics - metric indices for the graph, and 2) Random - metric indices for a random graph to compare against.
}
\description{
This function calculates a variety of indices for a network
}
\details{
A network graph is created for the network(s) that were generated.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
mynetwork <- network_calc(omicsData = rRNA_data)
mygraph <- pmartRseq_igraph(netData = mynetwork, coeff=0.6, pval=NULL, qval=0.05)
myindices <- network_indices(netGraph = mygraph)
}

}
\references{
Ju et al., 2014. Taxonomic relatedness shapes bacterial assembly in activated sludge of globally distributed wastewater treatment plants. Environ Microbiol. 2014 Aug;16(8):2421-32)
}
\author{
Allison Thompson
}
