% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{network_plot}
\alias{network_plot}
\title{Generate Network Plot}
\usage{
network_plot(omicsData, netData, coeff = NULL, pval = NULL, qval = NULL,
  colour = "Phylum", vsize = FALSE, legend.show = TRUE,
  legend.pos = "bottomleft")
}
\arguments{
\item{omicsData}{an object of the class 'seqData' usually created by \code{\link{as.seqData}}.}

\item{netData}{an object of class 'netData', created by \code{link{network_calc}}}

\item{coeff}{Optional, cutoff value to use for the correlation coefficient}

\item{pval}{Optional, cutoff value to use for p-values}

\item{qval}{Optional, cutoff value to use for q-values}

\item{colour}{Optional, if desired, can colour vertices by a taxonomic level}

\item{vsize}{Logical, should vertices be scaled by median abundance of taxa}

\item{legend.show}{Logical, should a legend be shown}

\item{legend.pos}{Optional, if legend==TRUE, where to position the legend. Default is 'bottomleft'.}
}
\value{
A network graph.
}
\description{
This function generates a network plot for the network data.
}
\details{
A network graph is created for the network(s) that were generated.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
mynetwork <- network_calc(omicsData = rRNA_data)
myNetworkPlot <- network_plot(mynetwork)
}

}
\author{
Allison Thompson
}
