% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(omicsData, norm_fn, normalize = FALSE, ...)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}.}

\item{norm_fn}{character vector indicating the normalization function to use for normalization. See details for valid options.}

\item{normalize}{For count data, this function will only return the scale and location parameters - will not return normalized data unless this parameter is set to TRUE. This is due to later statistics requiring raw data for count data analyses. Default is FALSE.}

\item{...}{additional arguments passed to the chosen normalization function.}
}
\value{
If normalize=FALSE, a list containing the location and scale parameters to use when normalizing the data. If normalize=TRUE, returns the omicsData object, where e_data has been normalized with the appropriate parameters and the scale and location parameters are returned as an attribute of the data.
}
\description{
Normalizes the data using the specified normalization function
}
\details{
For count data (16S data), the default normalization is Cumulative Sum Scaling \code{norm_fn="css"}.
The choices for normalization currently available, \code{norm_fn}, are:
\tabular{ll}{
 \code{"percentile"}  \tab Standardize the data by dividing each feature in \code{e_data} by the sample-wide qth percentile and multiply by the gloabal qth percentile\cr
 \tab \cr
 \code{"tss"}  \tab Standardize the data by dividing each feature in \code{e_data} by sum of the sample-wide counts\cr
 \tab \cr
 \code{"rarefy"} \tab Normalize the data by subsampling down to specified library size\cr
 \tab \cr
 \code{"poisson"} \tab Normalize the data by sampling from a Poisson distribution with appropriate mean value, see Section 2.2 of Li et al. (2013)\cr
 \tab \cr
 \code{"deseq"} \tab Normalization method used in DESeq and DESeq2, which uses size factors to standardize sequencing depths across samples\cr
  \tab \cr
 \code{"css"}  \tab Normalize the data by dividing each feature in \code{e_data} by the sum of the counts up to a specified quantile and multiplying by a global scaling factor\cr
 \tab \cr
 \code{"tmm"}  \tab Normalize the data using the trimmed mean of M values\cr
 \tab \cr
 \code{"none"} \tab No normalization is performed\cr
 \tab \cr
 }
}
\examples{
\dontrun{
library(mintJansson)

#Count data are passed to a quantile normalization by default
normalized_rRNAdata <- normalize_data(rRNA_data)
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = 'percentile')

#One can also use the TSS normalization for count data, the normalization function used by DESeq/DESeq2, normalization from SAMSeq (aka Poisson), another from metagenomeSeq (cumulative sum scaling normalization (CSS)), or yet another competitor edgeR (TMM normalization)
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = "tss")
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = "deseq")
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = "poisson")
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = "css")
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = "tmm")

#One could also rarefy the data - though this is highly NOT recommended
normalized_rRNAdata <- normalize_data(rRNA_data, norm_fn = "rarefy")
}

}
\references{
Li, Jun, and Robert Tibshirani. Finding consistent patterns: a nonparametric approach for identifying differential expression in RNA-Seq data. Statistical methods in medical research 22.5 (2013): 519-536.
 Anders, Simon and Wolfgang Huber. Differential expression analysis for sequence count data. Genome Biology 11:R106 (2010).
 Paulson, Joseph N, O Colin Stine, Hector Corrada Bravo, and Mihai Pop. Differential abundance analysis for microbial marker-gene surveys. Nature Methods. 10.12 (2013)
 Robinson, Mark D and Alicia Oshlack. A scaling normalization method for differential expression analysis of RNA-seq data. Genome Biology 11:R25 (2010).
}
\author{
Kelly Stratton, Lisa Bramer, Bryan Stanfill, Allison Thompson
}
