% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_diffabun.R
\name{plot_all_diffabun}
\alias{plot_all_diffabun}
\title{Plot all differential abundance results}
\usage{
plot_all_diffabun(countSTAT_results, omicsData, x_axis = "Phylum",
  facet = "~Comparison", scales = "fixed", x_lab = NULL, y_lab = NULL,
  leglab = NULL, plot_title = NULL)
}
\arguments{
\item{countSTAT_results}{Required, an object of the class \code{countSTAT} created by \code{\link{countSTAT}}.}

\item{omicsData}{Required, omicsData ,an object of the class \code{cDNAdata}, \code{gDNAdata}, or \code{rRNAdata} usually created by \code{\link{as.cDNAdata}}, \code{\link{as.gDNAdata}}, or \code{\link{as.rRNAdata}}, respectively.}

\item{x_axis}{Required, a character vector stating which variable to group data by and put on the x-axis, must be one of "Comparison" or one of the column names in omicsData$e_meta. Default is "Phylum".}

\item{facet}{Optional, a character vector depicting the formula to use when faceting the plots. Default is "~Comparison".}

\item{scales}{Optional, a character vector to describe if any of the axes should be free when faceting. Default is "fixed".}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{plot_title}{Optional, a character vector to use as the plot title}
}
\value{
A plot showing log2 fold change and p-values for all comparisons and all features
}
\description{
This function takes the differential abundance results of an omicsData object and creates a plot showing all the log2 fold changes of all features and the corresponding p-values. This looks at the entire dataset to help determine shifts in the whole community.
}
\examples{
\dontrun{
library(mintJansson)
library(mintR)
data("rRNAdata")


rRNAdata <- group_designation(omicsData=rRNA_data, main_effects="Treatment")
norm_factors <- normalize_data(omicsData=rRNAdata, subset_fn="none", norm_fn="percentile", normalize=FALSE)
norm_data <- normalize_data(omicsData=rRNAdata, subset_fn="none", norm_fn="percentile", normalize=TRUE)
diffexp <- countSTAT(omicsData=rRNAdata, norm_factors=norm_factors, comparisons="all", test="dw", pval_adjust="none", pval_thresh=0.05)

plot_all_diffabun(countSTAT_results=diffexp, omicsData=norm_data, x_axis="Phylum", facet="~Comparison", scales="fixed")
}

}
\author{
Allison Thompson
}
