% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indsp.R
\name{plot_indsp}
\alias{plot_indsp}
\title{Plot Indicator species Analysis Results}
\usage{
plot_indsp(indsp, omicsData, x_axis = "Group", group = "Phylum")
}
\arguments{
\item{indsp}{an indicator species analysis results object, created by \code{\link{indsp_calc}}}

\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}.}

\item{x_axis}{which grouping value to put on the x-axis. Default is "Group".}

\item{group}{Which taxonomic level to group the results by. Default is "Phylum".}
}
\value{
A plot of indicator species analysis results.
}
\description{
This function plots the results from an indicator species analysis.
}
\details{
This function creates a heatmap-type plot showing the results of an indicator species analysis.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
rRNA_data <- group_designation(omicsData = rRNA_data, main_effects = c("site", "treatment"), time_course=NULL)
rRNA_norm <- normalize_data(omicsData = rRNA_data, norm_fn = "css", normalize = TRUE)

rRNA_indsp <- indsp_calc(omicsData = rRNA_norm, within=NULL, pval_thresh=0.05)
head(rRNA_indsp)
summary(rRNA_indsp)
plot_indsp(indsp = rRNA_indsp, omicsData = rRNA_data)
}

}
\references{
De Cáceres, M. and Legendre, P. 2009. Associations between species and groups of sites: indices and statistical inference. Ecology 90(12): 3566-3574.
}
\author{
Allison Thompson
}
