% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pmartRseq.R
\name{plot_pmartRseq}
\alias{plot_pmartRseq}
\alias{plot.countSTAT_results}
\alias{plot_pmartRseq}
\alias{plot.alphaRes}
\alias{plot_pmartRseq}
\alias{plot.evenRes}
\alias{plot_pmartRseq}
\alias{plot.jaccardRes}
\alias{plot_pmartRseq}
\alias{plot.countFilter}
\alias{plot_pmartRseq}
\alias{plot.sampleFilter}
\alias{plot_pmartRseq}
\alias{plot.rRNAdata}
\alias{plot_pmartRseq}
\alias{plot.cDNAdata}
\alias{plot_pmartRseq}
\alias{plot.gDNAdata}
\alias{plot_pmartRseq}
\alias{plot.richRes}
\alias{plot_pmartRseq}
\alias{plot.abunRes}
\alias{plot_pmartRseq}
\alias{plot.effspRes}
\alias{plot_pmartRseq}
\alias{plot.indspRes}
\title{Produce a plot of an omicsData Object}
\usage{
plot.countSTAT_results(results_object, type = "pvals", test = NULL,
  x_lab = NULL, y_lab = NULL, plot_title = NULL, leglab = NULL,
  FCThresh = log2(1.5), comparison = NULL, ...)

plot.alphaRes(results_object, x_axis = "Group", color = "Group",
  shape = NULL, plot_title = NULL, scales = "free_y", x_lab = NULL,
  y_lab = NULL, leglab = NULL, ...)

plot.evenRes(results_object, x_axis = "Group", color = "Group",
  shape = NULL, plot_title = NULL, scales = "free_y", x_lab = NULL,
  y_lab = NULL, leglab = NULL, ...)

plot.jaccardRes(results_object, variable = "Median", x_axis = "Group",
  color = "Group", shape = NULL, plot_title = NULL, x_lab = NULL,
  y_lab = NULL, leglab = NULL, ...)

plot.countFilter(results_object, breaks = 100, max_count = NULL,
  min_num = NULL, min_samp = NULL, plot_title = NULL, x_lab = NULL,
  y_lab = NULL, ...)

plot.sampleFilter(results_object, breaks = 100, max_count = NULL,
  min_num = NULL, plot_title = NULL, x_lab = NULL, y_lab = NULL, ...)

plot.rRNAdata(results_object, x_axis = "Group", class = "Phylum",
  grp_fn = "median", subset_by = NULL, subset_val = NULL,
  plot_title = NULL, x_lab = NULL, y_lab = NULL, leglab = NULL, ...)

plot.cDNAdata(results_object, x_axis = "Group", class = "ECNum",
  grp_fn = "median", subset_by = NULL, subset_val = NULL,
  plot_title = NULL, x_lab = NULL, y_lab = NULL, leglab = NULL, ...)

plot.gDNAdata(results_object, x_axis = "Group", class = "ECNum",
  grp_fn = "median", subset_by = NULL, subset_val = NULL,
  plot_title = NULL, x_lab = NULL, y_lab = NULL, leglab = NULL, ...)

plot.richRes(results_object, abun = NULL, x_axis = "Group",
  color = "Group", scales = "free_y", shape = NULL, plot_title = NULL,
  x_lab = NULL, y_lab = NULL, leglab = NULL, ...)

plot.abunRes(results_object, rich = NULL, x_axis = "Group",
  color = "Group", shape = NULL, plot_title = NULL, x_lab = NULL,
  y_lab = NULL, leglab = NULL, ...)

plot.effspRes(results_object, x_axis = "Group", color = "Group",
  shape = NULL, plot_title = NULL, x_lab = NULL, y_lab = NULL,
  leglab = NULL, ...)

plot.indspRes(results_object, type = "pvals", ...)
}
\arguments{
\item{results_object}{is an object of class countSTAT_results (\code{\link{countSTAT}}), alphaDiversity (\code{\link{alphaDiversity}}), evenness (\code{\link{evenness}}), jaccard (\code{\link{mint_jaccard}}), countFilter (\code{\link{count_based_filter}}), rRNAdata (\code{\link{as.rRNAdata}}), cDNAdata (\code{\link{as.cDNAdata}}), gDNAdata (\code{\link{as.gDNAdata}}), richness (\code{\link{richness}}), abundance (\code{\link{abundance}}), effectiveSpecies (\code{\link{effectiveSpecies}}), or indicatorSpecies (\code{\link{indicatorSpecies}}).}

\item{type}{Character vector specifying which type of plot to create. "pvals" for a line plot showing the number of significantly expresed biomolecules at different p-value thresholds, "flag" for a bar plot showing the number of significantly expressed biomolecules at the previously specified threshold, "logfc" for a heatmap showing the log2 fold changes of the significantly expressed biomolecules, and/or "volcano" for a volcano plot showing the log2 fold changes versus -log10(pvalue). The default is "pvals".}

\item{test}{Optional, a character vector specifying which differential expression test to plot. The default is NULL, which will create plots for all tests.}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{FCThresh}{Optional, numeric threshold for the Log2FoldChanges for the volcano plot. Default is log2(1.5)}

\item{x_axis}{Required, a character vector specifying which variable to put on the x-axis, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{color}{Optional, a character vector specifying which variable to map to colors, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{shape}{Optional, a character vector specifying which variable to map to shape, must be one of the column names in attr(results_object, "group_DF"). Default is NULL.}

\item{scales}{Optional, a character vector to describe if any of the axes should be free when faceting. Default is "free_y" for a free y-axis.}

\item{variable}{Required, character vector specifying Which Jaccard variable to plot - options are "Median", "InterQuartileRange", "Average", and "StdDev". Default is "Median".}

\item{breaks}{Required, a number specifying the number of breaks to have in the cumulative graph. Default is 100.}

\item{max_count}{Optional, a number specifying the maximum count number to show on the graph. Default is NULL.}

\item{min_num}{Optional, a number specifying the desired cut point in order to visualize how many OTUs would be lost if that cut point was used. sum_based_filter uses strictly less than the desired min_num, so this will show the valus for strictly less than the desired min_num. Default is NULL.}

\item{min_samp}{Optional, for k/a filtering, a number specifying that OTUs must be seen in at least this many samples. Default is 2 for ka filters and NULL for everything else.}

\item{class}{Required, a character vector specifying which variable to group data by (e.g. "Phylum") and to color data by, must be one of the column names in results_object$e_data, results_object$e_meta, or attr(results_object, "group_DF"). Default is "Phylum".}

\item{grp_fn}{Required, a character vector specifying which function to use to summarise values across replicates. Can use one of "median", "mean", or "sum". Default is "median".}

\item{subset_by}{Optional, which variable to subset the data by. NULL will not subset the data. Default is NULL.}

\item{subset_val}{Optional, which values to include in the subset.}

\item{abun}{Optional, an object of class 'abundance', used for plotting richness vs abundance}

\item{rich}{Optional, an object of class 'richness', used for plotting richness vs abundance}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to put on the x-axis, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{color}{Optional, a character vector specifying which variable to map to colors, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{shape}{Optional, a character vector specifying which variable to map to shape, must be one of the column names in attr(results_object, "group_DF"). Default is NULL.}

\item{scales}{Optional, a character vector to describe if any of the axes should be free when faceting. Default is "free_y" for a free y-axis.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to put on the x-axis, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{color}{Optional, a character vector specifying which variable to map to colors, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{shape}{Optional, a character vector specifying which variable to map to shape, must be one of the column names in attr(results_object, "group_DF"). Default is NULL.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{breaks}{Required, a number specifying the number of breaks to have in
the cumulative graph. Default is 100.}

\item{max_count}{Optional, a number specifying the maximum count number to
show on the graph. Default is NULL.}

\item{min_num}{Optional, a number specifying the desired cut point in order to
visualize how many Sampless would be lost if that cut point was used.
sample_based_filter uses strictly less than the desired min_num, so this
will show the valus for strictly less than the desired min_num. Default is
NULL.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to group data by and put on the x-axis, must be one of the column names in results_object$e_data, results_object$e_meta, or attr(results_object, "group_DF"). Default is "Group".}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to group data by and put on the x-axis, must be one of the column names in results_object$e_data, results_object$e_meta, or attr(results_object, "group_DF"). Default is "Group".}

\item{class}{Required, a character vector specifying which variable to group and color data by, must be one of the column names in results_object$e_data, results_object$e_meta, or attr(results_object, "group_DF"). Default is "ECNum".}

\item{grp_fn}{Required, a character vector specifying which function to use to summarise values across replicates. Can use one of "median", "mean", or "sum". Default is "median".}

\item{subset_by}{Optional, which variable to subset the data by. NULL will not subset the data. Default is NULL.}

\item{subset_val}{Optional, which values to include in the subset.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to group data by and put on the x-axis, must be one of the column names in results_object$e_data, results_object$e_meta, or attr(results_object, "group_DF"). Default is "Group".}

\item{class}{Required, a character vector specifying which variable to group and color data by, must be one of the column names in results_object$e_data, results_object$e_meta, or attr(results_object, "group_DF"). Default is "ECNum".}

\item{grp_fn}{Required, a character vector specifying which function to use to summarise values across replicates. Can use one of "median", "mean", or "sum". Default is "median".}

\item{subset_by}{Optional, which variable to subset the data by. NULL will not subset the data. Default is NULL.}

\item{subset_val}{Optional, which values to include in the subset.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to put on the x-axis, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{color}{Optional, a character vector specifying which variable to map to colors, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{scales}{Optional, a character vector specifying if any/both of the axes be free. Default is "free_y", for a free y-axis.}

\item{shape}{Optional, a character vector specifying which variable to map to shape, must be one of the column names in attr(results_object, "group_DF"). Default is NULL.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to put on the x-axis, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{color}{Optional, a character vector specifying which variable to map to colors, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{shape}{Optional, a character vector specifying which variable to map to shape, must be one of the column names in attr(results_object, "group_DF"). Default is NULL.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{x_axis}{Required, a character vector specifying which variable to put on the x-axis, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{color}{Optional, a character vector specifying which variable to map to colors, must be one of the column names in attr(results_object, "group_DF"). Default is "Group".}

\item{shape}{Optional, a character vector specifying which variable to map to shape, must be one of the column names in attr(results_object, "group_DF"). Default is NULL.}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}

\item{type}{Required, a character vector specifying which type of plot to create. "pvals" for a line plot showing the number of significantly expresed biomolecules at different p-value thresholds; or "flag" for a bar plot showing the number of indicated species in each group, at the previously specified threshold. The default is "pvals".}

\item{plot_title}{Optional, a character vector to use as the plot title}

\item{leglab}{Optional, a character vector to use as the legend label}

\item{x_lab}{Optional, a character vector to use as the x-axis label}

\item{y_lab}{Optional, a character vector to use as the y-axis label}
}
\description{
This function will proivde a plot for a \code{omicsData} or \code{omicsData_results} object.
}
