% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq.R
\docType{package}
\name{pmartRseq}
\alias{pmartRseq}
\alias{pmartRseq-package}
\title{pmartRseq: A pipeline package that enables a reproducible workflow and
analysis of 16S data.}
\description{
The functions in this package are also avavilable as a
dashboard that allows for a complete analysis with no programming required.
There are four major components to this package: input and grouping,
filtering, normalization, and analysis.
}
