% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_aldex2.R
\name{pmartRseq_aldex2}
\alias{pmartRseq_aldex2}
\title{Modified version of ALDEx2 for pmartRseq}
\usage{
pmartRseq_aldex2(omicsData, mainEffects = NULL, mc.samples = 128,
  denom = "all", verbose = FALSE, interactions = FALSE,
  randomEffect = NULL, pval_thresh = 0.05)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' reated by \code{\link{as.seqData}}.}

\item{mainEffects}{Optional, a character vector detailing which factors should be used as main effects in the model. The variable name must match a column name from \code{omicsData$f_data}. If NULL, will use the same main effects that were used in \code{\link{group_designation}}.}

\item{mc.samples}{The number of Monte Carlo instances to use. Default is 128.}

\item{denom}{Character vector indicating which features to use as the denominator for the geometric mean. Default is "all".}

\item{verbose}{Print diagnostic information while running. Default is FALSE.}

\item{interactions}{Logical, indicating whether or not interactions should be considered in the model. If TRUE, all possible interactions will be used. If FALSE, none will be used.}

\item{randomEffect}{Optional, character specifying which factor to use as a random effect in the model. This will be used in the form (1|randomEffect). Must match a column name from \code{omicsData$f_data}}
}
\value{
An object of class paRes (also a list) containing to parts - clr, the clr transformed values for each Monte-Carlo Dirichlet instance, and res, a data frame with a p.value, statistic, degrees of freedom, and sum of squares values for each feature and each term in the mdoel.
}
\description{
This function calculates the centered-log ratio with draws from a Dirichlet distribution and runs a liner or linear mixed-effects model
}
\details{
Modified version of ALDEx2, allowing for multiple factors and random effects
}
\examples{
\dontrun{

}

}
\references{
Fernandes, Unifying...
}
\author{
Allison Thompson
}
