% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_DESeq2.R
\name{pmartRseq_DESeq2}
\alias{pmartRseq_DESeq2}
\title{DESeq2 analysis of omicsData objects}
\usage{
pmartRseq_DESeq2(omicsData, norm_factors = NULL, test = "wald", pairs, adj,
  thresh)
}
\arguments{
\item{omicsData}{an object of the class 'gDNAdata', 'cDNAdata', or 'rRNAdata' usually created by \code{\link{as.gDNAdata}}, \code{\link{as.cDNAdata}},or \code{\link{as.rRNAdata}}, respectively.}

\item{norm_factors}{Named vector of normalization parameters to put into DESeq2. If NULL, will use DESeq2's inbuilt normalization. Default is NULL.}

\item{test}{name of which differential expression test to use, options are "wald" or "paired". Default is "wald". See details for further explanation.}

\item{pairs}{a matrix dictating which pairwise comparisons to make}

\item{adj}{multiple comparison adjustment method to use.}

\item{thresh}{p-value threshold for significance.}
}
\value{
DESeqREsults object, which is a simple subclass of DataFrame. Columns include baseMean, log2FoldChange, lfcSE (standard error of log2FoldChange), stat (Wald statistic), pvalue, and padj (BH adjusted p-values). Use mcols(res)$description.
}
\description{
Differential abundance analysis of count data using DESeq2
}
\details{
Performs differential abundance testing on two groups using DESeq2.
}
\examples{
library(mintJansson)
data(cDNA_hiseq_data)
mycdnadata <- group_designation(omicsData = cDNA_hiseq_data, main_effects = c("treatment"), time_course=NULL)
mycdnadata_norm <- normalize_data(omicsData = mycdnadata, norm_fn = "percentile")
mycdnadata_DESeq2 <- pmartRseq_DESeq2(omicsData = mycdnadata_norm, test="wald", pairs = cbind(list("Neg","Plus")), adj = "BH", thresh = 0.05)

}
\references{
Love MI, Huber W and Anders S (2014). Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2. Genome Biology, 15, pp. 550.
}
\author{
Allison Thompson
}
