% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_igraph.R
\name{pmartRseq_igraph}
\alias{pmartRseq_igraph}
\title{Calculates a network graph}
\usage{
pmartRseq_igraph(netData, coeff = NULL, pval = NULL, qval = NULL)
}
\arguments{
\item{netData}{an object of class 'netData', created by \code{link{network_calc}}}

\item{coeff}{Optional, cutoff value to use for the correlation coefficient}

\item{pval}{Optional, cutoff value to use for p-values}

\item{qval}{Optional, cutoff value to use for q-values}
}
\value{
An object of class networkGraph (also a data.frame or list) containing the network.
}
\description{
This function calculates a network graph, using igraph, in order to create a network plot and/or calcuate network indices.
}
\details{
Create an igraph object, which can then be used to visualize a network plot and/or calculate network indices.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
mynetwork <- network_calc(omicsData = rRNA_data)
myNetGraph <- pmartRseq_igraph(mynetwork)
}

}
\author{
Allison Thompson
}
