% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_NMDS.R
\name{pmartRseq_NMDS}
\alias{pmartRseq_NMDS}
\title{NMDS plot for beta diversity}
\usage{
pmartRseq_NMDS(res, omicsData, grp, k, x_axis = "NMDS1", y_axis = "NMDS2",
  ellipses = TRUE)
}
\arguments{
\item{res}{an object created by vegan::metaMDS}

\item{omicsData}{a seqData object}

\item{grp}{name of column to use for grouping variables}

\item{k}{number of dimensions}

\item{x_axis}{which dimension to put on the x-axis}

\item{y_axis}{which dimension to put on the y-axis}

\item{ellipses}{logical indicating whether or not to show ellipses on the plot}
}
\value{
An NMDS plot of a beta diversity index.
}
\description{
This function creates an NMDS plot for a beta diversity object from vegan.
}
\details{
After beta diversity is calculated, this function will create a plot of the results, along with ellipses, to show the different groupings.
}
\examples{
\dontrun{
library(mintJansson)
library(vegan)
data(rRNA_data)
rRNA_data <- group_designation(omicsData = rRNA_data, main_effects = c("treatment"))
rRNA_norm <- normalize_data(omicsData = rRNA_data, norm_fn = "css", normalize = TRUE)
rRNA_veg <- pmartRseq_to_vegan(omicsData = rRNA_norm)
rRNA_metamds <- vegan::metaMDS(rRNA_veg, distance = "bray", k = 4, autotransform = FALSE, na.rm = TRUE)
pmartRseq_NMDS(rRNA_metamds, as.factor(attr(rRNA_norm,"group_DF")[match(rownames(rRNA_veg),attr(rRNA_norm,"group_DF")[,attr(rRNA_norm,"cnames")$fdata_cname]),"Group"]))
}

}
\author{
Allison Thompson
}
