% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_to_vegan.R
\name{pmartRseq_to_vegan}
\alias{pmartRseq_to_vegan}
\title{Transforms an omicsData object into a vegan data object}
\usage{
pmartRseq_to_vegan(omicsData)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}}
}
\value{
A data object for use with the package 'vegan'
}
\description{
This function takes an omicsData object and transforms it into a data object suitable for use in the vegan package
}
\examples{
\dontrun{

library(vegan)
library(mintJansson)
data(rRNA_data)

data <- pmartRseq_to_vegan(rRNA_data)
ord <- metaMDS(data)
plot(ord, display="sites", type="points")
}

}
\references{
Jari Oksanen, F. Guillaume Blanchet, Michael Friendly, Roeland Kindt, Pierre Legendre, Dan McGlinn, Peter R. Minchin, R. B. O'Hara, Gavin L. Simpson, Peter Solymos, M. Henry H. Stevens, Eduard Szoecs and Helene Wagner (2016). vegan: Community Ecology Package. R package version 2.4-1. https://CRAN.R-project.org/package=vegan
}
\author{
Allison Thompson
}
