% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{poisson_norm}
\alias{poisson_norm}
\title{Normalization of count data via Poisson Sampling}
\usage{
poisson_norm(e_data, edata_id)
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}
}
\value{
List containing 3 elements: norm_data is a data.frame with same structure as e_data that contains the normalized data, location_param is NULL, scale_param is a numeric vector of DESeq scores.
}
\description{
The method normalizes count data by resampling the data from a Poisson distribution with parameter estimated from the raw counts.  More specifically,
the quantity that is returned is given by equation 2.4 of Li and Tibshirani (2013).
}
\details{
Count data resampled from the Poisson distribution
}
\examples{
library(mintJansson)
e_data <- rRNA_data$e_data
e_data_id <- attr(rRNA_data, "cnames")$edata_cname
pois_ndata <- poisson_norm(rRNA_data$e_data, e_data_id)

}
\references{
Li, Jun, and Robert Tibshirani. "Finding consistent patterns: a nonparametric approach for identifying differential expression in RNA-Seq data." Statistical methods in medical research 22.5 (2013): 519-536.
}
\author{
Bryan Stanfill
}
