% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{Quant_Norm}
\alias{Quant_Norm}
\title{Quantile normalization of count data}
\usage{
Quant_Norm(e_data, edata_id, q = 0.75)
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}

\item{q}{a number to indicate which quantile to normalize with. Default is 0.75.}
}
\value{
List containing 3 elements: norm_data is a data.frame with same structure as e_data that contains the quantile-normalized data, location_param is NULL, and scale_param is a numeric vector containing, for every sample, the value of the sample at the designated quantile (q) divided by the value of the global quantile (q).
}
\description{
The method normalizes count data by a specified quantile (e.g., 75th quantile)
}
\details{
Count data is normalized by a given quantile, dividing by the given quantile of the sample and multiplying by the given quantile of the entire dataset.
}
\examples{
library(mintJansson)
data(cDNA_hiseq_data)
cDNA_quant <- Quant_Norm(e_data = cDNA_hiseq_data$e_data, edata_id = attr(cDNA_hiseq_data, "cnames")$edata_cname)
norm_factors <- attr(cDNA_quant,"data_info")$scale_param

}
\author{
Allison Thompson, Lisa Bramer
}
