% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{Rarefy}
\alias{Rarefy}
\title{Rarefying of count data}
\usage{
Rarefy(e_data, edata_id, size = NULL)
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}

\item{size}{the library size to rarefy down to. Default uses the minimum sample size.}
}
\value{
List containing 4 elements: norm_data is a data.frame with same structure as e_data that contains the rarefied data, location_param is NULL, and scale_param is the library size counts were rarefied to.
}
\description{
The method normalizes count data by rarefying
}
\details{
Count data is normalized by rarefying, subsampling samples down to a specified library size. If the specified library size is larger than a sample's library size, the sample will be discarded. A warning message will display which samples are discarded. This normalization method is likely not the best course of action.
}
\examples{
library(mintJansson)
data(cDNA_hiseq_data)
cDNA_rarefy <- Rarefy(e_data = cDNA_hiseq_data$e_data, edata_id = attr(cDNA_hiseq_data, "cnames")$edata_cname)
library_size <- attr(cDNA_rarefy,"data_info")$scale_param

}
\references{
McMurdie, Paul J., and Susan Holmes. "Waste Not, Want Not: Why Rarefying Microbiome Data Is Inadmissible." PLOS Computational Biology. 10.4 (2014)
}
\author{
Allison Thompson
}
