% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_calc.R
\name{richness_calc}
\alias{richness_calc}
\title{Calculates Richness}
\usage{
richness_calc(omicsData, index = c("observed", "chao1", "ace", "break"))
}
\arguments{
\item{omicsData}{an object of the class 'gDNAdata', 'cDNAdata', or 'rRNAdata' usually created by \code{\link{as.gDNAdata}}, \code{\link{as.cDNAdata}},or \code{\link{as.rRNAdata}}, respectively.}

\item{index}{a character vector stating which of the calculations to perform - "observed" for the observed richness, "chao1" the bias-corrected chao1 richness estimator, and/or "ace" for the abundance-based coverage richness estimator. Default is to perform all 3 calculations.}
}
\value{
An object of class richRes (also a data.frame) containing the richness value for every sample in the data object.
}
\description{
This function calculates the unique number of features seen in each sample.
}
\details{
Calculates richness of count data
}
\examples{
library(mintJansson)
data(rRNA_data)
rRNA_richness <- richness_calc(omicsData = rRNA_data)
rRNA_richness
summary(rRNA_richness)
plot(rRNA_richness)

}
\references{
Chao, Anne. Species Richness Estimation.
}
\author{
Allison Thompson
}
