% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_based_filter.R
\name{sample_based_filter}
\alias{sample_based_filter}
\title{Sample filter object}
\usage{
sample_based_filter(omicsData, fn = "sum")
}
\arguments{
\item{omicsData}{An object of the classes "seqData"}

\item{fn}{Specify "sum" to use the total OTU count of each sample and "criteria"
to use sample names}
}
\value{
An object of class sampleFilter (also a data.frame) that contains the
  sample identifier and the sum count across all OTUs.
}
\description{
This function returns a sampleFilter object, performing all of the
calculations needed to filter the data based off a specified function and
limit
}
\examples{

}
\author{
Allison Thompson and Sarah Reehl
}
