% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_emeta.R
\name{split_emeta}
\alias{split_emeta}
\title{Splits an e_meta column into separate columns}
\usage{
split_emeta(omicsData, cname = NULL, split1 = ";", numcol = NULL,
  split2 = "__", num = 2, newnames = NULL)
}
\arguments{
\item{omicsData}{an object of one of the class "seqData"}

\item{cname}{the column in e_meta to split. If NULL, will use attr(omicsData,"cnames")$emeta_cname}

\item{split1}{variable used to split string into columns. Default is ";".}

\item{numcol}{number of columns to split given column into. If NULL, will calculate from data. Default is NULL.}

\item{split2}{variable used to split string again, if needed. For example, can split "k__Bacteria" into "Bacteria" and will remove the "k__" part. Default is "__".}

\item{num}{which part to keep after splitting for the second time. Default is 2.}

\item{newnames}{new column names after splitting is done. If NULL, will use c("Kingdom","Phylum","Class","Order","Family","Genus","Species"). Default is NULL.}
}
\value{
seqData object with a split e_meta
}
\description{
This function can take one large string (like a taxonomy) and split it into a desired number of different columns. For instance, split a large string containing a taxonomy into columns for Kingdom, Phylum, etc.
}
\examples{
\dontrun{
library(mintJansson)
load(rRNA_data)

omicsData <- rRNA_data
head(omicsData$e_meta)

omicsData_split <- split_emeta(omicsData, emeta_cname="Taxonomy", split1=";", numcol=7, split2="__", num=2, newnames=NULL)
head(omicsData_split$e_meta)
}

}
\author{
Allison Thompson
}
