% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pmartRseq.R
\name{summary-pmartRseq}
\alias{summary-pmartRseq}
\alias{summary.seqData}
\alias{summary-pmartRseq}
\alias{summary.countSTAT_results}
\alias{summary-pmartRseq}
\alias{summary.alphaRes}
\alias{summary-pmartRseq}
\alias{summary.evenRes}
\alias{summary-pmartRseq}
\alias{summary.richRes}
\alias{summary-pmartRseq}
\alias{summary.abunRes}
\alias{summary-pmartRseq}
\alias{summary.jaccardRes}
\alias{summary-pmartRseq}
\alias{summary.effspRes}
\alias{summary-pmartRseq}
\alias{summary.countFilter}
\alias{summary-pmartRseq}
\alias{summary.indspRes}
\alias{summary-pmartRseq}
\alias{summary.paRes}
\title{Produce a basic summary of an pmartRseq S3 Object}
\usage{
\method{summary}{seqData}(omicsData)

\method{summary}{countSTAT_results}(pmartRseq_results)

\method{summary}{alphaRes}(pmartRseq_results)

\method{summary}{evenRes}(pmartRseq_results)

\method{summary}{richRes}(pmartRseq_results)

\method{summary}{abunRes}(pmartRseq_results)

\method{summary}{jaccardRes}(pmartRseq_results)

\method{summary}{effspRes}(pmartRseq_results)

\method{summary}{countFilter}(pmartRseq_results, min_num = NULL)

\method{summary}{indspRes}(pmartRseq_results)

\method{summary}{paRes}(pmartRseq_results)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}}

\item{min_num}{An integer value specifying the minimum number of fn counts a feature must have to be kept for statistical analysis. If fn="percent", give the decimal number, not the percentage.}
}
\value{
a list of elements
\itemize{
\item{num.samps} number of samples in \code{e_data}
\item{num.edata} number of unique identifying features in \code{e_data}
\item{num.emeta} number of unique identifying features in \code{e_meta}
\item{num_na} number of observations equal to NA in \code{e_data}
\item{frac_na} fraction of observations equal to NA in \code{e_data}
\item{num_zero} number of observations equal to 0 in \code{e_data}
\item{frac_zero} fraction of observations equal to 0 in \code{e_data}
}
}
\description{
This function will provide basic summary statistics about an object of the class \code{seqData}.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
summary(rRNA_data)
}

}
\author{
Lisa Bramer, Kelly Stratton, Allison Thompson
}
