% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_rollup.R
\name{taxa_rollup}
\alias{taxa_rollup}
\title{Roll up data to a specified taxonomic level}
\usage{
taxa_rollup(omicsData, level, taxa_levels = NULL)
}
\arguments{
\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}.}

\item{level}{taxonomic level to roll up to.}

\item{taxa_levels}{The levels of taxonomy (or other e_meta object) which might be used in the roll up. If NULL, will use c("Kingdom","Phylum","Class","Order","Family","Genus","Species"), in that order. Default is NULL.}
}
\value{
A seqData object of the same class as the input, where e_data and e_meta are rolled up to the specified level.
}
\description{
This function rolls up data to the specified taxonomic level so that statistics may be calculated at various taxonomic levels.
}
\details{
Data will be rolled (summed) up to a specified taxonomic level. For example, data at the OTU level could be rolled (summed) up to the Genus level before statistics are computed.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)

rRNA_split <- split_emeta(rRNA_data)

rRNA_rollup <- taxa_rollup(omicsData = rRNA_split, level = "Phylum")

dim(rRNA_rollup$e_data)
attributes(rRNA_rollup)
}

}
\author{
Allison Thompson
}
