% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{TMM_Norm}
\alias{TMM_Norm}
\title{Trimmed Mean of M Values normalization of count data}
\usage{
TMM_Norm(e_data, edata_id, reference = NULL, qm = 0.3, qa = 0.05)
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}

\item{reference}{which column in e_data should be used as the reference, default is to use the sample with the least amount of missing data.}

\item{qm}{percentage by which to trim M values (gene-wise log-fold-changes), default is 0.30 (30\%)}

\item{qa}{percentage by which to trim A values (absolute expression levels), default is 0.05 (5\%)}
}
\value{
List containing 3 elements: norm_data is a data.frame with same structure as e_data that contains the TMM-normalized data, location_param is a numeric vector of the TMM values for each sample, and scale_param is NULL.
}
\description{
The method normalizes count data by the trimmed mean of m values in each sample
}
\details{
Count data is normalized by the trimmed mean of m values.
}
\examples{
library(mintJansson)
data(cDNA_hiseq_data)
cDNA_TMM <- Quant_Norm(e_data = cDNA_hiseq_data$e_data, edata_id = attr(cDNA_hiseq_data, "cnames")$edata_cname)
norm_factors <- attr(cDNA_TMM,"data_info")$scale_param

}
\author{
Allison Thompson, Lisa Bramer
}
