% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funcs.R
\name{TSS_Norm}
\alias{TSS_Norm}
\title{Total sum scaling normalization of count data}
\usage{
TSS_Norm(e_data, edata_id)
}
\arguments{
\item{e_data}{a \eqn{p \times n} data.frame of count data, where \eqn{p} is the number of features and \eqn{n} is the number of samples. Each row corresponds to data for a feature, with the first column giving the feature name.}

\item{edata_id}{character string indicating the name of the feature identifier. Usually obtained by calling \code{attr(omicsData, "cnames")$edata_cname}.}
}
\value{
List containing 3 elements: norm_data is a data.frame with same structure as e_data that contains the TSS-normalized data, location_param is NULL, scale_param is a numeric vector of the total count in each sample.
}
\description{
The method normalizes count data by the total number of counts in each sample
}
\details{
Count data is normalized by the total count, dividing by the total count of each sample
}
\examples{
library(mintJansson)
data(cDNA_hiseq_data)
cDNA_TSS <- Quant_Norm(e_data = cDNA_hiseq_data$e_data, edata_id = attr(cDNA_hiseq_data, "cnames")$edata_cname)
norm_factors <- attr(cDNA_TSS,"data_info")$scale_param

}
\author{
Allison Thompson, Lisa Bramer
}
