% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{applyFilt}
\alias{applyFilt}
\alias{applyFilt}
\alias{applyFilt.countFilter}
\title{Apply a filter to a mintR S3 object}
\usage{
applyFilt(filter_object, omicsData, ...)

\method{applyFilt}{countFilter}(filter_object, omicsData, upper_lim = 2)
}
\arguments{
\item{filter_object}{an object of the class "countFilter", created by \code{count_based_filter}.}

\item{omicsData}{an object of the class \code{cDNAdata}, \code{gDNAdata}, or \code{rRNAdata} usually created by \code{\link{as.cDNAdata}}, \code{\link{as.gDNAdata}}, or \code{\link{as.rRNAdata}}, respectively.}

\item{upper_lim}{OTUs must have a max/mean/percent/nonmiss/sum number of counts above this threshold. OTUs with a count less than or equal to this number will be removed.}
}
\value{
An object of the class \code{cDNAdata}, \code{gDNAdata}, or \code{rRNAdata}, with specified cname_ids, edata_cnames, and emeta_cnames filtered out of the appropriate datasets.
}
\description{
This function takes a filter object of class "countFilter", and applies the filter to a dataset of class \code{cDNAdata}, \code{gDNAdata}, or \code{rRNAdata}.
}
\examples{
library(mintJansson)
data(rRNA_data)

to_filter <- count_based_filter(omicsData = rRNA_data, fn="mean")
rRNAdata2 <- applyFilt(filter_object = to_filter, omicsData = rRNA_data, upper_lim = 2)
print(str(attributes(rRNAdata2)$filters))

to_filter2 <- count_based_filter(omicsData = rRNAdata, fn="max")
rRNAdata3 <- applyFilt(filter_object = to_filter2, omicsData = rRNAdata, upper_lim = 2)
print(str(attributes(rRNAdata3)$filters))

}
\seealso{
\code{\link{count_based_filter}}
}
\author{
Allison Thompson
}
