% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{applyFilt_worker}
\alias{applyFilt_worker}
\title{Remove items that need to be filtered out}
\usage{
applyFilt_worker(filter_object, omicsData, group)
}
\arguments{
\item{filter_object}{a list created by the functions above}

\item{omicsData}{an object of the class 'seqData' created by \code{\link{as.seqData}}.}
}
\value{
list similar to as.omicsData object
}
\description{
This function removes filtered objects
}
\author{
Kelly Stratton, Lisa Bramer, Allison Thompson
}
