% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{network_plot}
\alias{network_plot}
\title{Generate Network Plot}
\usage{
network_plot(netGraph, omicsData = NULL, modData = NULL,
  colour = "Phylum", vsize = FALSE, legend.show = TRUE,
  legend.pos = "bottomleft")
}
\arguments{
\item{netGraph}{an object of class 'networkGraph', created by \code{\link{pmartRseq_igraph}}}

\item{omicsData}{Optional, an object of the class 'seqData' usually created by \code{\link{as.seqData}}, if want to colour by taxonomy and/or scale vertices by abundance}

\item{modData}{Optional, an object of class 'modData', created by \code{\link{detect_modules}}, if want to colour by modules.}

\item{colour}{Optional, if desired, can colour vertices by a taxonomic level or 'Module' for module. Use 'NULL' if no colour is desired.}

\item{vsize}{Logical, should vertices be scaled by median abundance of taxa}

\item{legend.show}{Logical, should a legend be shown}

\item{legend.pos}{Optional, if legend==TRUE, where to position the legend. Default is 'bottomleft'.}
}
\value{
A network graph.
}
\description{
This function generates a network plot for the network data.
}
\details{
A network graph is created for the network(s) that were generated.
}
\examples{
\dontrun{
library(mintJansson)
data(rRNA_data)
mynetwork <- network_calc(omicsData = rRNA_data)
mygraph <- pmartRseq_igraph(netData = mynetwork, coeff=0.6, pval=NULL, qval=0.05)
network_plot(omicsData = rRNA_data, netGraph = mygraph, colour = "Phylum", vsize = TRUE, legend.show = TRUE, legend.pos = "bottomleft")
}

}
\author{
Allison Thompson
}
