% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_NMDS.R
\name{pmartRseq_NMDS}
\alias{pmartRseq_NMDS}
\title{NMDS plot for beta diversity}
\usage{
pmartRseq_NMDS(res, grp)
}
\arguments{
\item{res}{an object created by vegan::metaMDS}

\item{grp}{vector of grouping variables}
}
\value{
An NMDS plot of a beta diversity index.
}
\description{
This function creates an NMDS plot for a beta diversity object from vegan.
}
\details{
After beta diversity is calculated, this function will create a plot of the results, along with ellipses, to show the different groupings.
}
\author{
Allison Thompson
}
