% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmartRseq_NMDS.R
\name{pmartRseq_NMDS}
\alias{pmartRseq_NMDS}
\title{NMDS plot for beta diversity}
\usage{
pmartRseq_NMDS(res, omicsData, grp, k, x_axis = "NMDS1", y_axis = "NMDS2",
  ellipses = TRUE)
}
\arguments{
\item{res}{an object created by vegan::metaMDS}

\item{omicsData}{a seqData object}

\item{grp}{name of column to use for grouping variables}

\item{k}{number of dimensions}

\item{x_axis}{which dimension to put on the x-axis}

\item{y_axis}{which dimension to put on the y-axis}

\item{ellipses}{logical indicating whether or not to show ellipses on the plot}
}
\value{
An NMDS plot of a beta diversity index.
}
\description{
This function creates an NMDS plot for a beta diversity object from vegan.
}
\details{
After beta diversity is calculated, this function will create a plot of the results, along with ellipses, to show the different groupings.
}
\author{
Allison Thompson
}
